requirements.txt
# Core RAG requirements
chromadb==0.4.22
sentence-transformers==2.2.2
PyPDF2==3.0.1
python-docx==1.1.0
google-generativeai==0.4.0
httpx==0.26.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0

# FastAPI & related
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
pydantic==2.5.3

# Streamlit
streamlit==1.31.0

# Clerk authentication & JWT
PyJWT==2.8.0
python-jose[cryptography]==3.3.0
authlib==1.3.0
requests==2.31.0
cryptography==42.0.2

# Additional utilities
loguru==0.7.2

# Python standard library (included in Python)
# os, pathlib, json, logging, functools, typing, asyncio, datetime
